com.utilties.async 1.3.0
- Split out classes into their own files
- Optimized addressable async handling
