+++ Utilities.Async/Packages/com.utilities.async/Runtime/Async/AwaiterExtensions.cs
@@ -23,11 +23,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 using System.Runtime.CompilerServices;
-using System.Runtime.ExceptionServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using UnityEngine;
@@ -150,7 +146,7 @@ namespace Utilities.Async
         public static SimpleCoroutineAwaiter<Object> GetAwaiter(this ResourceRequest instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter<Object>();
-            var enumerator = InstructionWrappers.ResourceRequest(awaiter, instruction);
+            var enumerator = ResourceRequest(awaiter, instruction);
             RunOnUnityScheduler(() => RunCoroutine(enumerator));
             return awaiter;
         }
@@ -160,7 +156,7 @@ namespace Utilities.Async
         public static SimpleCoroutineAwaiter<AssetBundle> GetAwaiter(this AssetBundleCreateRequest instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter<AssetBundle>();
-            var enumerator = InstructionWrappers.AssetBundleCreateRequest(awaiter, instruction);
+            var enumerator = AssetBundleCreateRequest(awaiter, instruction);
             RunOnUnityScheduler(() => RunCoroutine(enumerator));
             return awaiter;
         }
@@ -168,7 +164,7 @@ namespace Utilities.Async
         public static SimpleCoroutineAwaiter<Object> GetAwaiter(this AssetBundleRequest instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter<Object>();
-            var enumerator = InstructionWrappers.AssetBundleRequest(awaiter, instruction);
+            var enumerator = AssetBundleRequest(awaiter, instruction);
             RunOnUnityScheduler(() => RunCoroutine(enumerator));
             return awaiter;
         }
@@ -194,7 +190,7 @@ namespace Utilities.Async
         internal static SimpleCoroutineAwaiter GetAwaiterReturnVoid(object instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter();
-            var enumerator = InstructionWrappers.ReturnVoid(awaiter, instruction);
+            var enumerator = ReturnVoid(awaiter, instruction);
             RunOnUnityScheduler(() => RunCoroutine(enumerator));
             return awaiter;
         }
@@ -202,7 +198,7 @@ namespace Utilities.Async
         internal static SimpleCoroutineAwaiter<T> GetAwaiterReturnSelf<T>(T instruction)
         {
             var awaiter = new SimpleCoroutineAwaiter<T>();
-            var enumerator = InstructionWrappers.ReturnSelf(awaiter, instruction);
+            var enumerator = ReturnSelf(awaiter, instruction);
             RunOnUnityScheduler(() => RunCoroutine(enumerator));
             return awaiter;
         }
@@ -253,252 +249,40 @@ namespace Utilities.Async
         }
 
         private class CoroutineRunner : MonoBehaviour { }
-
-        /// <summary>
-        /// Processes Coroutine and notifies completion with result.
-        /// </summary>
-        /// <typeparam name="T">The result type.</typeparam>
-        public class SimpleCoroutineAwaiter<T> : INotifyCompletion
+        private static IEnumerator ReturnVoid(SimpleCoroutineAwaiter awaiter, object instruction)
         {
-            private Exception exception;
-            private Action continuation;
-            private T result;
-
-            public bool IsCompleted { get; private set; }
-
-            public T GetResult()
-            {
-                Debug.Assert(IsCompleted);
-
-                if (exception != null)
-                {
-                    ExceptionDispatchInfo.Capture(exception).Throw();
-                }
-
-                return result;
-            }
-
-            public void Complete(T taskResult, Exception e)
-            {
-                Debug.Assert(!IsCompleted);
-
-                IsCompleted = true;
-                exception = e;
-                result = taskResult;
-
-                // Always trigger the continuation on the unity thread
-                // when awaiting on unity yield instructions.
-                if (continuation != null)
-                {
-                    RunOnUnityScheduler(continuation);
-                }
-            }
-
-            void INotifyCompletion.OnCompleted(Action notifyContinuation)
-            {
-                Debug.Assert(continuation == null);
-                Debug.Assert(!IsCompleted);
-
-                continuation = notifyContinuation;
-            }
+            // For simple instructions we assume that they don't throw exceptions
+            yield return instruction;
+            awaiter.Complete();
         }
 
-        /// <summary>
-        /// Processes Coroutine and notifies completion.
-        /// </summary>
-        public class SimpleCoroutineAwaiter : INotifyCompletion
+        private static IEnumerator ReturnSelf<T>(SimpleCoroutineAwaiter<T> awaiter, T instruction)
         {
-            private Exception exception;
-            private Action continuation;
-
-            public bool IsCompleted { get; private set; }
-
-            public void GetResult()
-            {
-                Debug.Assert(IsCompleted);
-
-                if (exception != null)
-                {
-                    ExceptionDispatchInfo.Capture(exception).Throw();
-                }
-            }
-
-            public void Complete(Exception e)
-            {
-                Debug.Assert(!IsCompleted);
-
-                IsCompleted = true;
-                exception = e;
-
-                // Always trigger the continuation on the unity thread
-                // when awaiting on unity yield instructions.
-                if (continuation != null)
-                {
-                    RunOnUnityScheduler(continuation);
-                }
-            }
-
-            void INotifyCompletion.OnCompleted(Action notifyContinuation)
-            {
-                Debug.Assert(continuation == null);
-                Debug.Assert(!IsCompleted);
-
-                continuation = notifyContinuation;
-            }
+            yield return instruction;
+            awaiter.Complete(instruction);
         }
 
-        private class CoroutineWrapper<T>
+        private static IEnumerator ResourceRequest(SimpleCoroutineAwaiter<Object> awaiter, ResourceRequest instruction)
         {
-            private readonly SimpleCoroutineAwaiter<T> awaiter;
-            private readonly Stack<IEnumerator> processStack;
-
-            public CoroutineWrapper(IEnumerator coroutine, SimpleCoroutineAwaiter<T> awaiter)
-            {
-                processStack = new Stack<IEnumerator>();
-                processStack.Push(coroutine);
-                this.awaiter = awaiter;
-            }
-
-            public IEnumerator Run()
-            {
-                while (true)
-                {
-                    var topWorker = processStack.Peek();
-                    bool isDone;
-
-                    try
-                    {
-                        isDone = !topWorker.MoveNext();
-                    }
-                    catch (Exception e)
-                    {
-                        // The IEnumerators we have in the process stack do not tell us the
-                        // actual names of the coroutine methods but it does tell us the objects
-                        // that the IEnumerators are associated with, so we can at least try
-                        // adding that to the exception output
-                        var objectTrace = GenerateObjectTrace(processStack);
-                        awaiter.Complete(default, objectTrace.Any() ? new Exception(GenerateObjectTraceMessage(objectTrace), e) : e);
-
-                        yield break;
-                    }
-
-                    if (isDone)
-                    {
-                        processStack.Pop();
-
-                        if (processStack.Count == 0)
-                        {
-                            awaiter.Complete((T)topWorker.Current, null);
-                            yield break;
-                        }
-                    }
-
-                    // We could just yield return nested IEnumerator's here but we choose to do
-                    // our own handling here so that we can catch exceptions in nested coroutines
-                    // instead of just top level coroutine
-                    if (topWorker.Current is IEnumerator item)
-                    {
-                        processStack.Push(item);
-                    }
-                    else
-                    {
-                        // Return the current value to the unity engine so it can handle things like
-                        // WaitForSeconds, WaitToEndOfFrame, etc.
-                        yield return topWorker.Current;
-                    }
-                }
-            }
-
-            private static string GenerateObjectTraceMessage(List<Type> objTrace)
-            {
-                var result = new StringBuilder();
-
-                foreach (var objType in objTrace)
-                {
-                    if (result.Length != 0)
-                    {
-                        result.Append(" -> ");
-                    }
-
-                    result.Append(objType);
-                }
-
-                result.AppendLine();
-                return $"Unity Coroutine Object Trace: {result}";
-            }
-
-            private static List<Type> GenerateObjectTrace(IEnumerable<IEnumerator> enumerators)
-            {
-                var objTrace = new List<Type>();
-
-                foreach (var enumerator in enumerators)
-                {
-                    // NOTE: This only works with scripting engine 4.6
-                    // And could easily stop working with unity updates
-                    var field = enumerator.GetType().GetField("$this", BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
-
-                    if (field == null)
-                    {
-                        continue;
-                    }
-
-                    var obj = field.GetValue(enumerator);
-
-                    if (obj == null)
-                    {
-                        continue;
-                    }
-
-                    var objType = obj.GetType();
-
-                    if (!objTrace.Any() || objType != objTrace.Last())
-                    {
-                        objTrace.Add(objType);
-                    }
-                }
-
-                objTrace.Reverse();
-                return objTrace;
-            }
+            yield return instruction;
+            awaiter.Complete(instruction.asset);
         }
-
-        private static class InstructionWrappers
-        {
-            public static IEnumerator ReturnVoid(SimpleCoroutineAwaiter awaiter, object instruction)
-            {
-                // For simple instructions we assume that they don't throw exceptions
-                yield return instruction;
-                awaiter.Complete(null);
-            }
-
-            public static IEnumerator ReturnSelf<T>(SimpleCoroutineAwaiter<T> awaiter, T instruction)
-            {
-                yield return instruction;
-                awaiter.Complete(instruction, null);
-            }
-
-            public static IEnumerator ResourceRequest(SimpleCoroutineAwaiter<Object> awaiter, ResourceRequest instruction)
-            {
-                yield return instruction;
-                awaiter.Complete(instruction.asset, null);
-            }
 
 #if UNITY_ASSET_BUNDLES
 
-            public static IEnumerator AssetBundleCreateRequest(SimpleCoroutineAwaiter<AssetBundle> awaiter, AssetBundleCreateRequest instruction)
-            {
-                yield return instruction;
-                awaiter.Complete(instruction.assetBundle, null);
-            }
+        private static IEnumerator AssetBundleCreateRequest(SimpleCoroutineAwaiter<AssetBundle> awaiter, AssetBundleCreateRequest instruction)
+        {
+            yield return instruction;
+            awaiter.Complete(instruction.assetBundle);
+        }
 
-            public static IEnumerator AssetBundleRequest(SimpleCoroutineAwaiter<Object> awaiter, AssetBundleRequest instruction)
-            {
-                yield return instruction;
-                awaiter.Complete(instruction.asset, null);
-            }
+        private static IEnumerator AssetBundleRequest(SimpleCoroutineAwaiter<Object> awaiter, AssetBundleRequest instruction)
+        {
+            yield return instruction;
+            awaiter.Complete(instruction.asset);
+        }
 
 #endif // UNITY_ASSET_BUNDLES
 
-        }
     }
 }
