+++ Utilities.Async/Packages/com.utilities.async/Runtime/Async/SimpleCoroutineAwaiter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4e8e330a2a9c8548afe64b73bd81eea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
